requirements.txt
aioredis==1.3.1
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.1.0
autobahn==20.7.1
Automat==20.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.2
channels==2.4.0
channels-redis==3.0.1
chardet==4.0.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
daphne==2.5.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.3
django-braces==1.14.0
django-crispy-forms==1.11.2
django-js-asset==1.2.2
django-mptt==0.12.0
django-mptt-admin==2.1.0
django-oauth-toolkit==1.5.0
django-rest-framework-social-oauth2==1.1.0
django-rest-knox==4.1.0
django-service-objects==0.7.0
django-settings-export==1.2.1
django-sslserver==0.22
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-social-oauth2==1.0.9
heroku==0.1.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itypes==1.2.0
Jinja2==3.0.1
jwcrypto==0.8
lxml==4.6.3
MarkupSafe==2.0.1
msgpack==1.0.0
numpy==1.20.3
oauthlib==3.1.0
pandas==1.2.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==2.1.0
pyOpenSSL==19.1.0
python-dateutil==1.5
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rest-social-auth==6.0.1
service-identity==18.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.1
urllib3==1.26.4
websockets==9.0.2
whitenoise==5.2.0
zope.interface==5.1.0
